platform_kickstart/requirements.txt
kubernetes
pyyaml
